+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Remove secondary switch functionality from `Multilevel Switch CC`
 * Implement interview procedure for `Multilevel Switch CC`
 * Upgrade Prettier and ESLint to make use of the new TS 3.7 syntax
+* Update Multilevel Sensor definitions to latest specs and rename some sensor types
+* Move sensor type and scale definitions to JSON config files
 
 ## 1.6.0 (2019-10-23)
 * Implement Multi Channel Association CC and prefer it to Association CC if possible